jan dev

+++ src/DynamixelDevice.cpp
@@ -6,31 +6,28 @@
  */ 
 #include"DynamixelDevice.h"
 
-DynamixelDevice::DynamixelDevice(DynamixelID aID): 
- mID(aID), mStatusReturnLevel(255)
+DynamixelDevice::DynamixelDevice(DynamixelID aID)
+    : mID(aID), mStatusReturnLevel(255)
 {
-	mStatus=DYN_STATUS_OK;
-	if(mID==BROADCAST_ID)
-	{
-		mStatusReturnLevel=0;
+	mStatus = DYN_STATUS_OK;
+	if (mID == BROADCAST_ID) {
+		mStatusReturnLevel = 0;
 	}
 }
 
 DynamixelStatus DynamixelDevice::changeId(uint8_t id)
 {
 	DynamixelStatus result;
-	result=write(DYN_ADDRESS_ID, id);
-	if(result==DYN_STATUS_OK)
-	{
-		mID=id;
+	result = write(DYN_ADDRESS_ID, id);
+	if (result == DYN_STATUS_OK) {
+	    mID = id;
 	}
 	return result;
 }
 
 uint8_t DynamixelDevice::statusReturnLevel()
 {
-	if(mStatusReturnLevel==255)
-	{
+	if (mStatusReturnLevel == 255) {
 		init();
 	}
 	return mStatusReturnLevel;
@@ -39,8 +36,7 @@ uint8_t DynamixelDevice::statusReturnLevel()
 void DynamixelDevice::statusReturnLevel(uint8_t aSRL)
 {
 	write(DYN_ADDRESS_SRL, aSRL);
-	if(status()==DYN_STATUS_OK)
-	{
+	if (status() == DYN_STATUS_OK) {
 		mStatusReturnLevel=aSRL;
 	}
 }
@@ -61,9 +57,10 @@ uint8_t DynamixelDevice::firmware()
 
 void DynamixelDevice::communicationSpeed(uint32_t aSpeed)
 {
-	uint8_t value=2000000/aSpeed-1;
-	if(value!=0) // forbid 2MBd rate, because it is out of spec, and can be difficult to undo
-	{
+	uint8_t value = (uint8_t)(2000000 / aSpeed - 1);
+    // forbid 2MBd rate, because it is out of spec, 
+    // and can be difficult to undo
+	if (value != 0) {
 		write(DYN_ADDRESS_BAUDRATE, value);
 	}
 }
